+++ R/plugin.R
@@ -124,6 +124,9 @@ solve_LP <- function(x, control = list()) {
     rEnv <- rLScreateEnv()
     #Create LINDO model object
     rModel <- rLScreateModel(rEnv)
+    # Install the log callback before any data is loaded.  log_env must stay
+    # referenced until the model is deleted (see lindoapi_set_logfunc).
+    log_env <- lindoapi_set_logfunc(rEnv, rModel, control)
 
     nErr <- lindoapi_load_lp(x, rEnv, rModel)
 
@@ -325,6 +328,9 @@ solve_QP <- function(x, control = list()) {
     rEnv <- rLScreateEnv()
     #Create LINDO model object
     rModel <- rLScreateModel(rEnv)
+    # Install the log callback before any data is loaded.  log_env must stay
+    # referenced until the model is deleted (see lindoapi_set_logfunc).
+    log_env <- lindoapi_set_logfunc(rEnv, rModel, control)
     
     nErr <- lindoapi_load_qp(x, rEnv, rModel, control)
     row_perm <- attr(nErr, "row_perm")